requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==9.5.0
gunicorn==20.1.0
